+++ UI.py
@@ -1,8 +1,10 @@
 import module
 import read_todo
+from Task import Task
 
 
 def main():
+    current_task = Task()
     result = "done"
     while result != "end":
         today_todo = read_todo.display_all_task()
@@ -11,7 +13,7 @@ def main():
         else:
             print("开始计时！！")
         newinput = input("请输入指令")
-        result = module.main(newinput)
+        result = module.main(newinput,current_task)
 
 
 if __name__ == "__main__":
