+++ todo.csv
@@ -2,7 +2,7 @@ name,sub_tasks_count,start_time_date,start_time_time,repetition_gap,repetition_g
 数学分析,0,2022-05-31,04:30:42,1,d,30,10,True,False
 线性代数,0,2022-05-31,04:39:42,1,d,30,30,True,False
 机器学习,0,2022-05-31,04:39:42,1,d,30,30,True,False
-统计,0,2022-05-31,04:39:42,1,d,30,30,True,False
+统计,-5,2022-05-31,04:39:42,1,d,30,30,True,False
 随机过程,0,2022-05-31,04:39:42,1,d,30,30,True,False
 算法,0,2022-05-31,04:39:42,1,d,30,30,True,False
 pandas,0,2022-05-31,04:39:42,1,d,30,30,True,False
